flask/requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
python-dotenv==1.1.1
requests==2.32.5
telnetlib3==2.0.4
urllib3==2.5.0
Werkzeug==3.1.3
openpyxl==3.1.5
